add documentation, clean up two camera issues
